pyproject.toml
[project]
name = "sdr"
version = "0.1.0"
description = "RTL-SDR software-defined radio lab with MCP server"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.0.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pyrtlsdr>=0.3.0",
    "scipy>=1.17.1",
    "setuptools>=70,<81",
    "sounddevice>=0.5.5",
]
